Move Vec3Array to pc and rename to Vec3Slice
