+++ pc/registration/icp/correspondence_test.go
@@ -5,12 +5,12 @@ import (
 	"testing"
 
 	"github.com/seqsense/pcgol/mat"
+	"github.com/seqsense/pcgol/pc"
 	"github.com/seqsense/pcgol/pc/storage/kdtree"
-	"github.com/seqsense/pcgol/pc/testutil"
 )
 
 func TestNearestPointCorresponder(t *testing.T) {
-	base := testutil.Vec3Array{
+	base := pc.Vec3Slice{
 		mat.Vec3{4, 1, 0},
 		mat.Vec3{1, 1, 0},
 		mat.Vec3{8, 1, 1},
@@ -20,7 +20,7 @@ func TestNearestPointCorresponder(t *testing.T) {
 	kdt := kdtree.New(base)
 	corr := NewNearestPointCorresponder(kdt, 3)
 
-	targets := testutil.Vec3Array{
+	targets := pc.Vec3Slice{
 		mat.Vec3{8, 1, 1},  // nearest=2
 		mat.Vec3{-8, 1, 1}, // no corresponding point
 		mat.Vec3{2, 1, 0},  // nearest=1
